LPS-54860 move to module
