+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/query/BooleanQueryImpl.java
@@ -20,7 +20,6 @@ import com.liferay.portal.kernel.search.BooleanClauseOccur;
 import com.liferay.portal.kernel.search.BooleanClauseOccurImpl;
 import com.liferay.portal.kernel.search.ParseException;
 import com.liferay.portal.kernel.search.Query;
-import com.liferay.portal.search.lucene.QueryTranslator;
 
 import java.util.ArrayList;
 import java.util.List;
